List packages of architecture "all" in other architecture Package files.
A deb file with the architecture (arch) of "all" should be listed in the
Package.gz of all relevant architectures.

This change adds the following:
 - DebRepositoryConfiguration now has a way to configure which archs are
applicable to the "all" arch.
    There is no UI to edit this list, but editing
.BuildServer/config/deb-repositories.xml is possible to add/remove items
from the list.
 - Enhances DebPackageStore so that if a DebianPackage has an arch of
"all" then the package is added to the archs in the list of relevant
archs returned by the configuration.
 - Changes the DebDownloadController to request all Packages (including
those with arch as "all" when building/serving the Packages and
Packages.gz files.