Fix cascading in DebFileModel and DebPackageModel.
Previously, both models had CascadeType.ALL or DELETE, which meant that
they both tried to act as parents in the relationship.

This caused deletions of DebPackages caused deletions of DebFiles which
in turn caused more deletions of DebPackages. Therefore, removing a
repository cleared out all packages and files leaving a nearly empty DB.

A side effect of this change is that removing a repository does not
cascade down to deleting any DebFileModel rows which may no longer have
a package pointing at them. These DebFileModel rows will need to be
removed when they are deemed to be no longer used. I plan to execute
this task after TeamCity cleanup has run.

Here are the main changes in the commit:

 - Made the relevant changes to the models and then ran MainDbMigration
which generates the new EBean migration scripts.

 - Added DebRepositoryMaintenanceManager interface and added methods to
DebRepositoryManagerImpl to clean up dangling DebFileModel rows (and
their DebPackageParameterModel children). This is not currently
configured to run in TeamCity, but the unit tests are working.
 
 - Added extra data to the dd7824da repo file, so assist with testing.
 
 - Added methods to access package and file counts (linked and dangling)
to present this information in admin tab.
 
 - Created a file to keep useful queries in.